Rename _typing to typing
