+++ transactron/utils/__init__.py
@@ -1,5 +1,5 @@
 from .data_repr import *  # noqa: F401
-from ._typing import *  # noqa: F401
+from .typing import *  # noqa: F401
 from .debug_signals import *  # noqa: F401
 from .assign import *  # noqa: F401
 from .amaranth_ext import *  # noqa: F401
