+++ transactron/utils/data_repr.py
@@ -1,5 +1,5 @@
 from collections.abc import Iterable, Mapping
-from ._typing import MethodLayout
+from .typing import MethodLayout
 from amaranth_types import ShapeLike
 from typing import Any, Hashable, Sized
 from statistics import fmean
